Restructure the files and added admin functions
