+++ server/models.py
@@ -38,6 +38,7 @@ class RedFlag(db.Model):
     video_file = db.Column(db.String, nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     date = db.Column(db.DateTime, default=datetime.utcnow)
+    status = db.Column(db.String(20), default='pending')  # 'pending', 'investigating', 'rejected', 'resolved'
     category = db.Column(db.String(255), nullable=False) #fix this
     admin_id = db.Column(db.Integer, db.ForeignKey('admin.id'), nullable=False)
 
@@ -55,17 +56,13 @@ class Intervention(db.Model):
     video_file = db.Column(db.String, nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     date = db.Column(db.DateTime, default=datetime.utcnow)
+    status = db.Column(db.String(20), default='pending' ) # 'pending', 'investigating', 'rejected', 'resolved'
     admin_id = db.Column(db.Integer, db.ForeignKey('admin.id'), nullable=False)
 
-class Admin(db.Model, UserMixin):
-    __tablename__ = 'admin'
-
+class AdminAction(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(10), unique=True, nullable=False)
-    # first_name = db.Column(db.String(50), nullable=False)
-    # last_name = db.Column(db.String(50), nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    password = db.Column(db.String(60), nullable=False)
-
-    red_flags = db.relationship('RedFlag', backref='user', lazy=True)
-    interventions = db.relationship('Intervention', backref='user', lazy=True)
+    post_id = db.Column(db.Integer, nullable=False)
+    post_type = db.Column(db.String, nullable=False)
+    status = db.Column(db.String, nullable=False)
+    admin_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
